Separate point table into own class.
